change dependencies
removed the microsoft gsl headers, using gsl-lite instead now.

+++ include/solvers.h
@@ -3,7 +3,7 @@
 #include "structs.h"
 
 namespace solvers {
-double effective_elasticity(double e1, double e2);
+float32 effective_elasticity(float32 e1, float32 e2);
 std::tuple<V2, V2> velocities_after_collision(
     const Physical_properties& p1, const V2& u1, 
     const Physical_properties& p2, const V2& u2);
